Use HTTPS for repository URI in docker-build.sh
For some (unknown) reason using plain HTTP for `quay.io` with the Accept
header times out (although without the header it works?!). Switching to
HTTPS seems to work with the header? It's all very confusing.

Also added `--show-error` to `curl` and only redirect standard out,
so that we get any errors logged.