Use a script for pushing images to multiple repos
While I was testing #112 I noticed that the images that are pushed to
Docker Hub were tagged with `v<date>-<sha>` instead of `pull-XXX`. I
remember @damdo having some issues with getting the `&&` to work
properly so I moved the logic for pushing to multiple repositories to a
script instead.